# Performance: cache PurgeCSS output