- Extract cache to its own file
- Create new `digest` util (uses Node.js crypto SHA256 -> hex hash digest)
- add caching for PurgeCSS output
- adds `size` measurement for Purged CSS (opens the path to maxSize support for CSS)

### Results

Uncached:

```sh
yarn example:uncached
yarn run v1.21.1
$ rm -rf .remote-asset-cache && yarn example
$ rm -rf dist && src/cli.js 'fixtures/*.html' -o dist
fixtures/sample.html -> dist/sample.html: 1831.639ms
✨  Done in 3.12s.
```

Cache

```sh
yarn example
yarn run v1.21.1
$ rm -rf dist && src/cli.js 'fixtures/*.html' -o dist
fixtures/sample.html -> dist/sample.html: 16.493ms
✨  Done in 0.63s.
```